+++ src/main/java/com/donkeys_today/server/support/exception/NotFoundException.java
@@ -1,6 +1,6 @@
-package com.donkeys_today.server.common.exception;
+package com.donkeys_today.server.support.exception;
 
-import com.donkeys_today.server.common.dto.type.ErrorType;
+import com.donkeys_today.server.support.dto.type.ErrorType;
 
 public class NotFoundException extends BusinessException{
